+++ src/pages/CartPages.tsx
@@ -1,57 +1,35 @@
+import { Suspense } from 'react';
 import { styled } from 'styled-components';
 import Header from '../components/common/Header/Header';
 import CartList from '../components/cart/CartList/CartList';
 import PaymentAmount from '../components/cart/PaymentAmount/PaymentAmount';
 import CheckedCartListProvider from '../provider/CheckedListProvider';
-import useCartService from '../hooks/useCartService';
-import { useNavigate } from 'react-router-dom';
-import { useEffect } from 'react';
-import { useRecoilValue } from 'recoil';
-import cartLoadingState from '../globalState/atoms/cartLoadingState';
-import serverNameState from '../globalState/atoms/serverName';
 import LoadingSpinner from '../components/common/LoadingSpinner/LoadingSpinner';
 
 const CartPage = () => {
-  const navigate = useNavigate();
-  const { cartList, fetchCartItem } = useCartService();
-  const isCartLoading = useRecoilValue(cartLoadingState);
-  const serverName = useRecoilValue(serverNameState);
-
-  const handleLinkButtonClick = () => {
-    navigate('/');
-  };
-
-  useEffect(() => {
-    fetchCartItem();
-  }, [serverName]);
-
   return (
-    <CheckedCartListProvider>
+    <>
       <Header />
       <Layout>
         <Title>장바구니</Title>
-        {isCartLoading ? (
-          <EmptyCartView>
-            <LoadingSpinner color="#06c09e" />
-          </EmptyCartView>
-        ) : cartList.length ? (
-          <Contents>
-            <>
+        <Suspense fallback={<LoadingView />}>
+          <CheckedCartListProvider>
+            <Contents>
               <CartList />
               <PaymentAmount />
-            </>
-          </Contents>
-        ) : (
-          <EmptyCartView>
-            장바구니에 상품이 존재하지 않습니다.
-            <LinkButton onClick={handleLinkButtonClick}>
-              상품 담으러 가기
-            </LinkButton>
-          </EmptyCartView>
-        )}
-        {/* </Suspense> */}
+            </Contents>
+          </CheckedCartListProvider>
+        </Suspense>
       </Layout>
-    </CheckedCartListProvider>
+    </>
+  );
+};
+
+const LoadingView = () => {
+  return (
+    <LoadingViewWrapper>
+      <LoadingSpinner color="#06c09e" />
+    </LoadingViewWrapper>
   );
 };
 
@@ -89,30 +67,12 @@ const Contents = styled.div`
   }
 `;
 
-const EmptyCartView = styled.div`
+const LoadingViewWrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap: 20px;
 
   margin-top: 200px;
-
-  font-weight: 500;
-  font-size: 30px;
-`;
-
-const LinkButton = styled.button`
-  width: 300px;
-  padding: 20px 50px;
-  background-color: #333;
-
-  border: none;
-  border-radius: 15px;
-
-  font-size: 20px;
-  color: white;
-
-  cursor: pointer;
 `;
 
 export default CartPage;
